Week4/Code_demo/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-swagger-ui==4.11.1
pyyaml==6.0.1
